Lengthy Commit - Please see README
This essentially creates Django App that can modularly add on data
stories via a CSV File we can write the stories to (make sure to inject
via a Script). This also lets us use Python for backend dataanalysis,
and then allows the backend data analysis to come forth as charts in a
django-javascript json connection.

Please see readme - it's made to be modular and to allow for future
volunteers to make a new data story (assuming they have the dataset/or
parse data already) with just writting some text in a csv and making a
few code changes (configuration like changes).

I am moving back to campus this weekend, but this was a fun project and
I hope it can be of help to you guys! Best of luck and I am incredibly
thankful for the work you guys are doing. I will likely talk to didi one
more time about how to use this as well to explain the nuances of this
commit.